Fix issue in create new version without selecting as default version

+++ components/apimgt/org.wso2.carbon.apimgt.gateway/src/main/java/org/wso2/carbon/apimgt/gateway/service/APIGatewayAdmin.java
@@ -798,7 +798,8 @@ public class APIGatewayAdmin extends org.wso2.carbon.core.AbstractAdmin {
         // Delete Default API
         String qualifiedDefaultApiName = GatewayUtils.getQualifiedDefaultApiName(gatewayAPIDTO.getProvider(),
                 gatewayAPIDTO.getName());
-        if (restapiAdminServiceProxy.getApi(qualifiedDefaultApiName) != null) {
+        if (restapiAdminServiceProxy.getApi(qualifiedDefaultApiName) != null && gatewayAPIDTO
+                .getDefaultAPIDefinition() != null) {
             restapiAdminServiceProxy.deleteApi(qualifiedDefaultApiName);
         }
         if (log.isDebugEnabled()) {
